fix: explicitly mark MainView as a CDI bean (#512)
Adding the `@CdiComponent` annotation to MainView allows it to be correctly
detected as a CDI bean, so that lifecycle hooks are invoked.
This is required because starting with Jakarta 10 bean-discovery-mode default
value is 'annotated'. Another solution may have been to modify beans.xml to set
bean-discovery-mode to 'all', but this is not the recommended way.